Small refactor and starting mobile dev
