+++ src/components/Appendix/ReaderPage/Books/DesktopBook/DesktopBook.tsx
@@ -3,7 +3,10 @@ import Styles from "./DesktopBook.module.scss";
 import TitleCoverSmall from "./TitleCover/TitleCoverSmall";
 import { EpubDetails } from "../../../../../interfaces/epub/epub-reader.interface";
 import { useEffect, useReducer } from "react";
-import { BookHolder, EpubReaderType, ReaderState, readerReducer } from "../helpers";
+import { BookHolder } from "../Helpers/Components/BookHolder";
+import { EpubReaderType } from "../Helpers/enums";
+import { ReaderState } from "../Helpers/interfaces";
+import { readerReducer } from "../Helpers/functions";
 import TitleCoverWide from "./TitleCover/TitleCoverWide";
 import dynamic from "next/dynamic";
 import { PageProps } from "../Page/Page";
