+++ src/components/Appendix/ReaderPage/Books/Helpers/Components/BookHolder.tsx
@@ -1,0 +1,8 @@
+import { styled, Box } from '@mui/material';
+
+export const BookHolder = styled(Box)(({
+    maxWidth: '100%',
+    maxHeight: '100%',
+    height: '100vh',
+    overflow: 'hidden'
+}));
