+++ src/components/Appendix/ReaderPage/Books/Helpers/enums.tsx
@@ -1,0 +1,11 @@
+export enum Orientiation {
+    landscape,
+    portrait
+}
+
+export enum EpubReaderType {
+    fullWidth,
+    halfPageWidth,
+    fullPageWidth
+}
+
