+++ src/components/Appendix/ReaderPage/Books/Helpers/functions.tsx
@@ -1,42 +1,12 @@
-import { styled } from '@mui/material';
-import { EpubHeader } from '../../../../interfaces/epub/epub-reader.interface';
-import { Box } from '@mui/system';
+import { EpubHeader } from "../../../../../interfaces/epub/epub-reader.interface";
+import { EpubReaderType } from "./enums";
+import { ReaderDisplayAction, ReaderState, SocialLinkDetails } from "./interfaces";
 import {
-    IconDefinition,
     faFacebookF,
     faRedditAlien,
     faTumblr,
     faTwitter,
 } from "@fortawesome/free-brands-svg-icons";
-
-export enum Orientiation {
-    landscape,
-    portrait
-}
-
-export enum EpubReaderType {
-    fullWidth,
-    halfPageWidth,
-    fullPageWidth
-}
-
-export interface ReaderDisplayAction {
-    type: EpubReaderType
-}
-
-export interface ReaderState {
-    largeEpub: string;
-    smallEpub: string;
-    currentScript: string;
-    setting: EpubReaderType;
-}
-
-export interface SocialLinkDetails {
-    key: string;
-    link: string;
-    label: string;
-    icon: IconDefinition;
-}
 
 export function readerReducer(state: ReaderState, action: ReaderDisplayAction) {
     const { type } = action;
@@ -78,10 +48,3 @@ export function coverSocialLinks(props: EpubHeader) : SocialLinkDetails[]{
         }
     ] as SocialLinkDetails[];
 }
-
-export const BookHolder = styled(Box)(({
-    maxWidth: '100%',
-    maxHeight: '100%',
-    height: '100vh',
-    overflow: 'hidden'
-}));
