+++ src/components/Appendix/ReaderPage/Books/TouchScreenBook/TouchScreenBook.module.scss
@@ -1,1 +1,3 @@
-
+.BookArea {
+    overflow-y: auto;
+}
